Repaired timestamp bug and added spinner for choosing rate.
The timestamp provided by the SensorEvent is actually the number of nanoseconds of uptime.
This, strangely, does not apply to the Location object, which has a normal epoch millisecond timestamp.

Also the UI is now equiped with a new spinner in the MainActivity in which the user can select the Rate of capture for all sensors.